pyproject.toml
[project]
name = "travel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "langchain==0.3.0",
  "langgraph==0.2.23",
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "httpx==0.27.2",
  "pydantic==2.9.2",
  "python-dotenv==1.0.1",
  "langchain-openai==0.2.0",
  "streamlit==1.39.0",
  "requests==2.32.3"
]